// frontend/src/api/orderService/model/orderResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { OrderResponseStatus } from './orderResponseStatus';
import type { OrderItemDto } from './orderItemDto';
import type { Address } from './address';

export interface OrderResponse {
  id?: string;
  customerId?: string;
  status?: OrderResponseStatus;
  totalAmount?: number;
  items?: OrderItemDto[];
  deliveryAddress?: Address;
  longitude?: number;
  latitude?: number;
  createdAt?: string;
  updatedAt?: string;
}